requirements.txt
fastapi
requests
python-dotenv
uvicorn
python-jose